chore: refactor PanelMenu
