+++ src/Components/ServiceScene/LogsVolumePanel.tsx
@@ -11,7 +11,7 @@ import { toggleLevelFromFilter } from 'services/levels';
 import { LoadingState } from '@grafana/data';
 import { getFieldsVariable, getLabelsVariable, getLevelsVariable } from '../../services/variableGetters';
 import { areArraysEqual } from '../../services/comparison';
-import { ExploreLogsVizPanelMenu, getPanelWrapperStyles } from '../Panels/ExploreLogsVizPanelMenu';
+import { PanelMenu, getPanelWrapperStyles } from '../Panels/PanelMenu';
 
 export interface LogsVolumePanelState extends SceneObjectState {
   panel?: VizPanel;
@@ -56,7 +56,7 @@ export class LogsVolumePanel extends SceneObjectBase<LogsVolumePanelState> {
       .setTitle('Log volume')
       .setOption('legend', { showLegend: true, calcs: ['sum'], displayMode: LegendDisplayMode.List })
       .setUnit('short')
-      .setMenu(new ExploreLogsVizPanelMenu({}))
+      .setMenu(new PanelMenu({}))
       .setData(
         getQueryRunner([
           buildDataQuery(getTimeSeriesExpr(this, LEVEL_VARIABLE_VALUE, false), {
