+++ src/services/fields.ts
@@ -26,7 +26,7 @@ import { getDetectedFieldsFrame } from '../Components/ServiceScene/ServiceScene'
 import { getLogsStreamSelector, getValueFromFieldsFilter } from './variableGetters';
 import { LabelType } from './fieldsTypes';
 import { logger } from './logger';
-import { ExploreLogsVizPanelMenu } from '../Components/Panels/ExploreLogsVizPanelMenu';
+import { PanelMenu } from '../Components/Panels/PanelMenu';
 
 export type DetectedLabel = {
   label: string;
@@ -148,7 +148,7 @@ export function getFilterBreakdownValueScene(
       )
       .setColor({ mode: 'fixed', fixedColor: getColorByIndex(frameIndex) })
       .setOverrides(setLevelColorOverrides)
-      .setMenu(new ExploreLogsVizPanelMenu({ frame, fieldName: getTitle(frame), labelName: labelKey }))
+      .setMenu(new PanelMenu({ frame, fieldName: getTitle(frame), labelName: labelKey }))
       .setHeaderActions([new AddToFiltersButton({ frame, variableName })]);
 
     if (style === DrawStyle.Bars) {
